Added Latent Space Animation
Model was trained for 30 epochs on a dataset consiting of all knowledge spaces of domain sizes 2 to 5. 
The embeddings derived from the input projections of all the surmise relations from our dataset are visualized using pca, tsne and umap dimensionality reduction techniques.
Red color indicates, that the given state (see plot title) is included in the knowledge structure represented by that data point.  
The size of the data points reflects the size of the knowledge structures, which were generated from the embeddings.